templates/flask-api/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
python-dotenv==1.0.1
gunicorn==23.0.0
